readme.md + seeder env change + adding some commands to package

+++ src/seeders/book.seeder.ts
@@ -1,6 +1,7 @@
 import { Seeder } from 'typeorm-extension';
 import { DataSource } from 'typeorm';
 import { Book } from '../entity';
+import env from '../env';
 
 export class BookSeeder implements Seeder {
   /**
@@ -11,24 +12,26 @@ export class BookSeeder implements Seeder {
   track = false;
 
   public async run(dataSource: DataSource): Promise<void> {
-    const repository = dataSource.getRepository(Book);
-    await repository.delete({});
-    await repository.insert([
-      {
-        name: '1984',
-      },
-      {
-        name: 'Brave New World',
-      },
-      {
-        name: 'Dune',
-      },
-      {
-        name: 'I, Robot',
-      },
-      {
-        name: "The Hitchhiker's Guide to the Galaxy",
-      },
-    ]);
+    if (env.APP.NODE_ENV === 'development') {
+      const repository = dataSource.getRepository(Book);
+      await repository.delete({});
+      await repository.insert([
+        {
+          name: '1984',
+        },
+        {
+          name: 'Brave New World',
+        },
+        {
+          name: 'Dune',
+        },
+        {
+          name: 'I, Robot',
+        },
+        {
+          name: "The Hitchhiker's Guide to the Galaxy",
+        },
+      ]);
+    }
   }
 }
